requirements.txt
torch==1.13.1
Pillow==9.4.0
numpy==1.24.2